+++ tccloud/__init__.py
@@ -1,5 +1,5 @@
 """TeraChem Cloud python client. Scalable compute, easy to learn, fast to code."""
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 from .client import TCClient  # noqa