+++ docs/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.3.1]
 
 ### Added
 
@@ -76,7 +76,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - `_RequestsClient` class that handles all network requests to TeraChem Cloud server
 - `FutureResults` object that is created from a `task_id` and can be used to retrieve a result once finished.
 
-[unreleased]: https://github.com/mtzgroup/tccloud/compare/0.3.0...HEAD
+[unreleased]: https://github.com/mtzgroup/tccloud/compare/0.3.1...HEAD
+[0.3.1]: https://github.com/mtzgroup/tccloud/releases/tag/0.3.1
 [0.3.0]: https://github.com/mtzgroup/tccloud/releases/tag/0.3.0
 [0.2.4]: https://github.com/mtzgroup/tccloud/releases/tag/0.2.4
 [0.2.3]: https://github.com/mtzgroup/tccloud/releases/tag/0.2.3